SWCForecast

Documentation for SWCForecast.

SWCForecast.FeatureCounts

SWCForecast.InfoShiftedFeature

SWCForecast.MyPath

SWCForecast.SubMD

SWCForecast.SubMD

Base.merge!

Base.merge!

Base.merge!

SWCForecast._create_index_html

SWCForecast._extend

SWCForecast._group_feature

SWCForecast._series2supervised

SWCForecast.addcol_accumulation!

SWCForecast.blankaxis!

SWCForecast.cccount

SWCForecast.checkparse

SWCForecast.chkdatetime

SWCForecast.chknnm

SWCForecast.compareresult

SWCForecast.convert_types

SWCForecast.convertdf2!

SWCForecast.copyimg

SWCForecast.copysections

SWCForecast.cropwhite

SWCForecast.datenum

SWCForecast.datenum

SWCForecast.datetimeticks!

SWCForecast.datetimeticks!

SWCForecast.describeinstr

SWCForecast.description!

SWCForecast.description!

SWCForecast.description!

SWCForecast.description!

SWCForecast.description!

SWCForecast.description!

SWCForecast.dict2mdliststr

SWCForecast.diffsstable!

SWCForecast.dropmissingcol

SWCForecast.elwmerge!

SWCForecast.emptyMD

SWCForecast.expandylim!

SWCForecast.feature_summary

SWCForecast.feature_summary

SWCForecast.featureselectbyheadkey

SWCForecast.filecolumnview

SWCForecast.filecolumnview

SWCForecast.findtrees

SWCForecast.forcontourf

SWCForecast.format_time_tag

SWCForecast.formatcolumn

SWCForecast.generatemyslide

SWCForecast.get1var

SWCForecast.get1var

SWCForecast.get1var

SWCForecast.getallfeat

SWCForecast.getallfeat

SWCForecast.getparam_atom

SWCForecast.getsection

SWCForecast.getxylimits

SWCForecast.group_feature

SWCForecast.ifstrparse

SWCForecast.imputemean!

SWCForecast.isableparse

SWCForecast.iscommented

SWCForecast.iscommentedand

SWCForecast.islevel

SWCForecast.islevelleq

SWCForecast.isvarexist

SWCForecast.iswhite

SWCForecast.latextable

SWCForecast.latextable

SWCForecast.latextable

SWCForecast.level_text

SWCForecast.load_decision_tree_regressor

SWCForecast.manyscript

SWCForecast.markdownreport

SWCForecast.markdownreport

SWCForecast.markdownreport

SWCForecast.mdimgpath!

SWCForecast.mdimgpath!

SWCForecast.mdimgpath!

SWCForecast.mdimgpath!

SWCForecast.mdimgpath!

SWCForecast.mdpaths

SWCForecast.mdpaths

SWCForecast.mdreport

SWCForecast.mkresultdir

SWCForecast.mvmean

SWCForecast.mvnanmean

SWCForecast.mycolor_makie_palette

SWCForecast.mycolor_plots_palette

SWCForecast.mymovwinmean

SWCForecast.myslideshow

SWCForecast.narrow_types!

SWCForecast.onlycodelines

SWCForecast.print2string

SWCForecast.quantileband!

SWCForecast.quantileband!

SWCForecast.quickdataoverview

SWCForecast.readdescription

SWCForecast.readdescription!

SWCForecast.recursive_merge

SWCForecast.recursive_merge!

SWCForecast.reduce_feature

SWCForecast.reducetype

SWCForecast.refbuildpath

SWCForecast.reline!

SWCForecast.reline!

SWCForecast.reline!

SWCForecast.replacerewrite

SWCForecast.result_folder_expr

SWCForecast.returnchild

SWCForecast.returnchild

SWCForecast.runallscript

SWCForecast.secondyaxis

SWCForecast.selectname

SWCForecast.selectnames

SWCForecast.series2supervised

SWCForecast.shrinkylim!

SWCForecast.split_time_tag

SWCForecast.symmquantile

SWCForecast.targetrange

SWCForecast.targetsection

SWCForecast.toordinal

SWCForecast.toordinal

SWCForecast.tree_structure

SWCForecast.tree_structure

SWCForecast.tree_structure

SWCForecast.treeinspect!

SWCForecast.writedescription

SWCForecast.FeatureCounts — Type

FeatureCounts of: maxlevel: the max level of nodes to inspect featids: the obtained identity number for features at nodes featvals: the obtained values of thresholds at nodes atlevels: the current level

Example

FC = FeatureCounts(4)
treeinspect!(FC, WEE1) # WEE1 is the DecisionTree.Root object

See also

treeinspect!

SWCForecast.InfoShiftedFeature — Type

Example

featstrs = [v for v in values(dmachs)][1][1].mach.data[1] |> names
ISFs = [InfoShiftedFeature(gf) for gf in group_feature(rmprefix.(featstrs))]

SWCForecast.MyPath — Type

a type that should be a path

SWCForecast.SubMD — Method

Given the path to the markdown file and a string headermatch, SubMD(path, headermatch) returns a SubMD object.

Fields

md::Markdown.MD
header::Vector{AbstractString} # e.g., `["## Introduction"]`
hash4::Vector{AbstractString} # e.g., `["c4a3"]`
path::Vector{AbstractString} # e.g. `"/training/decisiontree_20220309/RESULT_ExpeDeci_0x15b03b01222fec37/brief_report.md"`

SWCForecast.SubMD — Method

SubMD() create an empty SubMD object.

Base.merge! — Method

SWCForecast.merge!(md0::Markdown.MD, mds::Markdown.MD...)

Base.merge! — Method

SWCForecast.merge!(md0::Markdown.MD, contents::Vector...)

Example

md1 = md"""
# MyTitle with **Bold** string
A paragraph starts here
"""

contents = [
    Markdown.Header{1}(["Comparison"]),
    Markdown.Paragraph(
        ["This is the comparision of the section ",
        Markdown.Bold(md1.content[1].text), # `MyTitle with **Bold** string`
        " between results."])
]

preface = merge!(md"", contents)

which returns the following

Comparison

This is the comparision of the section MyTitle with Bold string between results.

Base.merge! — Method

For smd::SubMD, SWCForecast.merge!(smd0::SubMD, smds::SubMD...) merge (using push!) the smd.md.content,smd.header, smd.hash4, smd.path into smd0.

All smd in smds

Example

combinedmd = merge!(SubMD(), smds...)

SWCForecast._create_index_html — Method

This is modified from Remark._create_index_html

SWCForecast._extend — Method

Given the indices w0, w1 indicating the left- and right- non-white boundary of the image, returns the extended indices w0e and w1e by factor cpad_w.

SWCForecast._group_feature — Method

Given a vector of string str0 with each element being like "r_o_o_t_suffix", SWCForecast._group_feature(str0) returns the vector of (root_string, suffix_string).

SWCForecast._series2supervised — Method

series2supervised(data, range_shift, range_out)

Tansform a time series dataset into a supervised learning dataset.

The features will always be suffixed by an addtional time shift tag "t-i". Also see `splittimetag()andformattime_tag`.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

SWCForecast.addcol_accumulation! — Method

Add columns that are derived by accumulating corresponding7 variables.

Example

all_precipstr = names(df, r"precipitation")

apd = Dict( # time intervals to accumulates precipitation
    "1hour" => 6,
    "12hour" => 6*12,
    "1day" => 6*24,
    "2day" => 6*24*2,
    "3day" => 6*24*3
)

addcol_accumulation!(df, all_precipstr, apd)

SWCForecast.blankaxis! — Method

blankaxis!(fpos, xmin::T, ymin::T, xmax::T, ymax::T) where {T<:AbstractFloat} creates an empty Makie Axis object at grid position fpos.

Example

f = Figure(;resolution=(1400,1000))
gleft = f[1:3, 1] = GridLayout()
gright = f[1:3, 2:3] = GridLayout()
fpos = gright[0, :]
blankaxis!(fpos, xmin, ymin, xmax, ymax)

SWCForecast.cccount — Method

Of a time series ts, cccount(ts) calculate by default the cumulative counts of elements that approximates zero consecutively.

SWCForecast.checkparse — Method

checkparse(df::DataFrame, TYPE) check if all elements column variables can be parsed to a specific data TYPE. It returns a dictionary with the column names of df as its keys, and a vector of DataType and anything that cannot be parsed to TYPE.

SWCForecast.chkdatetime — Method

chkdatetime(v...) use try ... catch ... to check if a vector [yyyy, mm, dd, hh, MM, ss] is a valid datetime. It returns false if it is not an legal date vector array.

SWCForecast.chknnm — Method

chknnm(df) check if DataFrame df contains missing values or NaN. Use this before input df into machine.

SWCForecast.compareresult — Method

Given a vector strvec the last few letters (e.g., ["ac23", "7df6"]) of the hashes, compareresult(headerstr::AbstractString, strvec::Union{Vector{<:AbstractString}, Regex}, this_dir; fnameexpr=r"brief\_report\.md") will collect specific sections of the markdown files in the folders that are suffixed by letters in strvec (e.g., "RESULT_..._0x...7df6","RESULT...0x...ac23"), with section header matchingheaderstr, and export a comparison markdown file inthis_dir`.

Noted that strvec can be a regular expression such as r"RESULT"; see mdpaths for how it searces the target markdown files.

SWCForecast.convert_types — Method

convert_types(df, column_names_types)

Converts the element type of each column to a user-specified type.

Arguments

df: Dataframe for which you want to convert the eltype of each column

column_names_types: Column names and target types. The type of column_names_types should be able to be unpacked into column names and target types in a for loop.

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/11

SWCForecast.convertdf2! — Method

convertdf2!(df, TYPE) convert all elements in df to the type of TYPE, ignoring NaN and missing.

SWCForecast.copyimg — Method

Copy images that are referenced in the markdown script md to destinations.

SWCForecast.copysections — Method

copysections(headerstr::AbstractString, strvec::Union{Vector{<:AbstractString}, Regex}, result_dir, presentationdir::Function; fnameexpr=expr_defaultmdname, oneslide=false, renameheader="", appendhash=false) reads "brief_report.md" file in the target folders whose folder name matches strvec, and returns a vector Markdown.MD objects whose url for local images are relocated. This function is intended to return a vector of Markdown.MD objects with each element being one "slide" for Remark.js.

Noted that all images references by target files (those who name "brief_report.md") are copied to presentationdir("src", "assets", "img"), and the referenced paths to the images are relocated to "assets/img/...".

Keyword arguments

oneslide: strvec given as regular expression, only the first slide is returned if oneslide=true. This is useful when the target section headerstr of all files in result_dir is roughly identical.

renameheader: Set md.content[1].text as renameheader. For example, , renameheader=["Hello ", Markdown.Bold(["World"])]

Example

Get the "Brief summary" sections in the markdown files "brief_report.md" across all r"RESULT" folders, concatenated as slides (separated by ---):

headerstr = "## Brief Summary"
folderpattern = r"RESULT"
mds = copysections(headerstr, folderpattern, result_dir, presentationdir; fnameexpr="brief_report.md") |> join

A Literate example (one slide); the oneslide=true option returns only the first.

# ---
# ## Description

headerstr = "### Features" #hide
copysections(headerstr, r"RESULT", result_dir, presentationdir;oneslide=true)[1] #hide

# ---

A Literate example that returns slides, where there are total two slides with each page having the "### Data overview" and "### Prediction" sections:

# ---

strvec = ["052c", "9a2c"] #hide
headerstr = "### Data overview" #hide
mds = copysections(headerstr, strvec, resultdir0, presentationdir) #hide
headerstr = "### Prediction" #hide
mds2 = copysections(headerstr, strvec, resultdir0, presentationdir; appendhash=true) #hide
elwmerge!(mds, mds2) #hide
join(mds) #hide

# ---

Tips

The usage is similar to compareresult, but returns a vector mds::Vector{Markdown.MD} instead of one combined Markdown.MD. Use join(mds) to combine mds into one single Markdown.MD, separated by "---".

Use elwmerge! to combine the results elementwisely; this is useful if you want to have sections combined into one slide.

Also see Remark.jl.

SWCForecast.cropwhite — Method

Given the file path filei, cropwhite(filei::AbstractString; padding_w = 0.5, padding_h = 0.2, saveInplace=false) returns the cropped image

SWCForecast.datenum — Method

datenum(v...) does datenum(DateTime(v...)).

SWCForecast.datenum — Method

datenum(d::Dates.DateTime) converts a Julia DateTime to a MATLAB style DateNumber. Also see: toordinal

SWCForecast.datetimeticks! — Method

Similar to datetimeticks!(ax2, t::Vector{DateTime}, x_a::Vector), datetimeticks!(ax2, t::Vector{DateTime}, x_a::Vector, tinc::DatePeriod; datestrformat = "yyyy/mm/dd", modify_fn = identity) return xticks, xticklabels with xticks forced to increase with step tinc, and be modify by function modify_fn.

Example

datetimeticks!(
    ax2,
    [DateTime(2012,2,5,3,15,0), DateTime(2012,3,5,3,15,0)],
    [0, 1], # `x` for `Makie.plot(x, ...)`
    Day(3); # tick every 3 days
    datestrformat = "yyyy/mm/dd",
    modify_fn = x -> floor.(x, Day)
)

SWCForecast.datetimeticks! — Method

datetimeticks!(ax2, t::Vector{DateTime}, x::Vector; datestrformat = "yyyy/mm/dd") set x ticks to datestr format. t is the DateTime array that is not supported by Makie, x is a arbitrarily defined series of numbers that corresponds to t for Makie.plot. x and t must be the same length and should be pairwisely mapped.

Example:

t = df.datetime
x = Dates.datetime2epochms.(t)
x1 = x .- x[1] # to avoid glitchs in plotting with CairoMakie (due to the too-large values)
y = df.soil_water
CairoMakie.scatter!(ax, x, y, markersize =3)
datetimeticks!(ax2, t,x)

SWCForecast.describeinstr — Method

Given either missing, NaN and Number, returns "Missing", "Not a Number" and "Number".

SWCForecast.description! — Method

description!(DDT::AbstractDescription, str::AbstractString) update D.description with string str merged to it.

Example

str = "
    [database]
    server = "192.168.1.1"
    ports = [ 8001, 8001, 8002 ]
    ";
description!(DDT, str::AbstractString)

SWCForecast.description! — Method

description!(D::AbstractDescription, d::Dict) update D.description with d merged to it.

SWCForecast.description! — Method

Given the training machine mach, description!(DDT::DescriptCompTree, mach::Machine) update DDT.description with model information.

SWCForecast.description! — Method

Given the training machine mach, description!(DDT::DescriptEnsembleTrees, mach::Machine) update DDT.description with model information.

SWCForecast.description! — Method

Given the training machine mach, description!(DDT::DescriptOneForest, mach::Machine) update DDT.description with model information.

SWCForecast.description! — Method

Given the training machine mach, description!(DDT::DescriptOneTree, mach::Machine) update DDT.description with model information.

SWCForecast.dict2mdliststr — Method

Example

fitp = fitted_params(mach)
best_fitted_param_model = fitp.best_fitted_params.model # which should be a `Dict`

dict2mdliststr(best_fitted_param_model)

SWCForecast.diffsstable! — Method

Of a variable of name varnm, diffsstable!(X0::DataFrame, varnm, tshift) calculates the difference between the non-shifted (suffixed by "_t0") and time-shifted (e.g., "_t-6"), where the difference is the new column for the series-to-supervised table X0.

Example

    (X0,) = series2supervised(...)
    diffsstable!(X0, "precipitation_1hr", -6)

that creates a new column diff6_precipitation_1hr = X0[:, "precipitation_1hr_t0"] .- X0[:, "precipitation_1hr_t-6"].

SWCForecast.dropmissingcol — Method

dropmissingcol(df) drops columns which are all missing.

SWCForecast.elwmerge! — Method

Merge mds1 and mds2 elementwisely.

SWCForecast.emptyMD — Method

Simply return a empty Markdown.MD object.

SWCForecast.expandylim! — Method

expandylim!(ax, upper_expand) expand the upper limit of y axis to a ratio of upper_expand of the original y range (ymax - ymin).

Example

expandylim!(ax, 0.25)

SWCForecast.feature_summary — Method

Example

DDTdescript = readdescription(path2toml)
feats0 = DDTdescript["Model"]["features"]
targs0 = DDTdescript["Model"]["targets"]
list_features, list_targets = feature_summary(feats0, targs0)

SWCForecast.feature_summary — Method

feats0 = [
"pressure_CWB_t0" ,
"pressure_CWB_t-2" ,
"pressure_CWB_t-4" ,
"pressure_CWB_t-6" ,
"pressure_CWB_t-12" ,
"humidity_CWB_t0" ,
"humidity_CWB_t-2" ,
"humidity_CWB_t-4" ,
"humidity_CWB_t-6" ,
"humidity_CWB_t-12" ,
]

julia> feature_summary(feats0)
"- pressure CWB ($t_i, i=-12,-6,...,0$; total 5 variables) 
 - humidity CWB ($t_i, i=-12,-6,...,0$; total 5 variables)"

SWCForecast.featureselectbyheadkey — Method

Given feature names of the original table (before time shifted), featureselectbyheadkey(Xtrain, featsets0) returns the table where only variables with column name matches headkey_set in featsets0 are preserved. Noted that headkey in headkey_set must be the first keyword of the column name.

Example

Xtrain = DataFrame(
    "tmp" => randn(10),
    "tmp_t0" => randn(10),
    "tmp_t-1" => randn(10),
    "humidity" => randn(10),
    "humidity_t0" => randn(10),
    "pressure_t0" => randn(10),
    "pressure_t1" => randn(10),
)

featsets0 = [
    [:tmp, :humidity],
    [:tmp, :pressure],
    # comma is required even when there is only one union
]

and

julia> featureselectbyheadkey(Xtrain, featsets0)
2-element Vector{Vector{Symbol}}:
 [:tmp, :tmp_t0, Symbol("tmp_t-1"), :humidity, :humidity_t0]
 [:tmp, :tmp_t0, Symbol("tmp_t-1"), :pressure_t0, :pressure_t1]

SWCForecast.filecolumnview — Method

Providing the paths to csv files, filecolumnview(paths) gives a plot indicating the available columns of each file.

SWCForecast.filecolumnview — Method

Given a directory dir, filecolumnview(fexpr::Regex, dir) gives a plot indicating the available columns in files under dir that matches fexpr.

SWCForecast.findtrees — Method

Given a vector of machines, findtrees(machs::Vector{Machine}) returns a vector of only Machine{DecisionTreeRegressor, ...}.

SWCForecast.forcontourf — Method

forcontourf(allx::Vector, ally::Vector, xedges, yedges) returns xs, ys, zs for CairoMakie.contourf!(xs, ys, zs).

Noted that length(xs) + 1 = length(xedges).

Input argument

allx, ally are each a vector of values corrsponding to the X and Y coordinates in the plot, indicating all points on the X-Y plane.

xedges, yedges are each a range specifying the edges for calculating 2D histogram.

SWCForecast.format_time_tag — Method

format_time_tag format time tag into LaTeX maths.

Example

julia> format_time_tag(["t-1", "t-2", "t-3", "t-4"])
"t_{i=-4,-3,...,-1}"

SWCForecast.formatcolumn — Method

Format DataFrame's column if the element type belongs AbstractString:

add "\text{}"

add "\" to any dash ("_" => "\_") to prevent error in the LaTeX building stage

SWCForecast.generatemyslide — Method

Similar to Remark.generate(target_dir), generatemyslide(target_dir) generate the slide in the folder target_dir but with my template index.jl and style.css.

Tips and Tricks

set image width in "src/style.css" at img {max-width: ...}

SWCForecast.get1var — Method

get1var(df::DataFrame, expr::Regex) indexes df on the only one column that matches expr. If more than one columns are matched, it raises an error.

SWCForecast.get1var — Method

get1var(df::DataFrame) returns the only one column of df as a Vector. If there are more than one column, it raises an error.

SWCForecast.get1var — Method

get1var(v::Vector) get the only variable in the vector v which should have exactly one element; an error will be raised if length(v) != 1.

SWCForecast.getallfeat — Method

Get all features (a vector of column names).

SWCForecast.getallfeat — Method

Get all features (column names). Returns file names (ks) and columnames of each file (vls).

SWCForecast.getparam_atom — Method

Retrieve the atom model and learned parameters for each atoms

all_machs = fitted_params(mach).machines
findparam(mach::Machine{T, N}) where {T, N} = T

tps = all_machs .|> findparam # datatypes
targetmodelnamepattern = r"MLJ.*DecisionTreeRegressor"
# which matches:
# "MLJEnsembles.DeterministicEnsembleModel{DecisionTreeRegressor}"
# "MLJTuning.DeterministicTunedModel{RandomSearch, MLJEnsembles.DeterministicEnsembleModel{DecisionTreeRegressor}}"
idisforest = occursin.(targetmodelnamepattern, string.(tps))
forests_trained = all_machs[idisforest]

atomeachforest = getmodel_atom.(forests_trained)
parameachforest = getparam_atom.(forests_trained)
# [f.fitresult.fitresult.atom for f in forests_trained]

treemodels = []
treestructs = []
for (atom, treestruct) in zip(atomeachforest, parameachforest)
    push!(treestructs, treestruct)
    push!(treemodels, fill(atom, length(treestruct)))
end

learnedparams_tree = vcat(treestructs...)
atoms = vcat(treemodels...)
ntrees = length(learnedparams_tree)
# 1st model is TunedModel; 2nd is EnsembleModel

Also see

[7] build_forest(labels::AbstractVector{T}, features::AbstractMatrix{S}, n_subfeatures, n_trees, partial_sampling, max_depth, min_samples_leaf, min_samples_split) where {S, T<:Float64} in DecisionTree at /.../DecisionTree/iWCbW/src/regression/main.jl:49

fit!(rf::RandomForestRegressor, X::AbstractMatrix, y::AbstractVector) at /.../DecisionTree/iWCbW/src/scikitlearnAPI.jl:300

WrappedEnsemble(atom, ensemble::AbstractVector{L}) where L at /.../MLJEnsembles/OsgHR/src/ensembles.jl

MMI.fitted_params(::RandomForestRegressor, forest) = (forest=forest,) and

MMI.predict(::RandomForestRegressor, forest, Xnew) at /.../MLJDecisionTreeInterface/CtTJy/src/MLJDecisionTreeInterface.jl:255

SWCForecast.getsection — Method

Give the path of the markdown file, returns a Markdown.MD object of the section headerstr.

Example

path = "/training/brief_report.md"
headerstr = "## Description"
getsection(path, headerstr)

In this example, the section starts from "## Description" and ends right before the next level 2 (e.g., "## blablabla") or level 1 (e.g., "# blabla") header; comments in code fences won't be recognized as a header.

Also see SubMD.

SWCForecast.getxylimits — Method

Example

xmin, ymin, xmax, ymax = getxylimits(ax)

SWCForecast.group_feature — Method

Given a vector of string str0 with each element being like "r_o_o_t_suffix", group_feature(str0) returns an Iterator with each a subgroup of elements having the same root.

SWCForecast.ifstrparse — Method

ifstrparse(x, TYPE): if x is AbstractString, then parse it to the variable of TYPE, ignoring NaN and missing.

SWCForecast.imputemean! — Method

imputemean!(df) substitute missing values with the statistical means. If all missing for a column, value 999 is imputed.

Notice

You should be aware that imputemean! might does nothing without error message if the input is a view of dataframe (e.g., df[!, Not(:datetime)]).

SWCForecast.isableparse — Method

isableparse(x, TYPE) use try...catch to check if x can be parsed to of TYPE. If not, return x; else, return TYPE. If x::TYPE it also returns TYPE evenif it can't be parsed to the type of itself.

SWCForecast.iscommented — Method

iscommented(oneline::AbstractString) returns true if the line is commented by "#".

SWCForecast.iscommentedand — Method

iscommentedand(oneline::AbstractString, tag) returns true if the line is commented and starts with tag.

SWCForecast.islevel — Method

Given a object mdc, islevel(mdc, n) returns true if mdc is the type of Markdown.Header{n}

SWCForecast.islevelleq — Method

Given a object mdc, islevelleq(mdc, n) returns true if mdc is the type of Markdown.Header{x} where x ≤ n. In brief, it recursively finds if it is a header of higher level (smaller n) until n==0 (false is returned).

SWCForecast.isvarexist — Method

isvarexist(paths) returns dataframes (df1, df2) in the "long" format with each row indicating whether a variable in a file exists. In which, (df, dfi) both have columns [:varname, :fname, :isexist]. In df, all variables are CategoricalArrays; in dfi, all variables are integers for indexing levels(df.varname), levels(df.fname) and levels(df.isexist).

SWCForecast.iswhite — Method

Check if the pixel is white or transparent.

SWCForecast.latextable — Method

latextable(dftmp, MIME("text/latex")) returns a very simple latex table that applies MIME("text/latex"). Other MIME is of-coursely unsupported. This method is experimental following Mimes In Julia: What Are They?.

SWCForecast.latextable — Method

Other Support:

ISFs = [InfoShiftedFeature(gf) for gf in group_feature(names(X0))]
latextable(ISFs::Vector{InfoShiftedFeature})

which is equivalent to

dftmp = DataFrame("feature name" => String[], "time shift" => String[])
for ISF1 in ISFs
    push!(dftmp, [ISF1.root, join(string.(ISF1.timeshift), " ,")])
end
latextable(dftmp::DataFrame)

SWCForecast.latextable — Method

latextable(manytables::Vector{Pair{String,DataFrame}}) is under construction. Noted that

SWCForecast.level_text — Method

Example

level_text("## Introduction") returns the header level and the appending string (2, "Introduction").

SWCForecast.load_decision_tree_regressor — Method

load_decision_tree_regressor(; package="DecisionTree", kwargs...)

A wrapper function for MLJ @load macro that returns a DecisionTreeRegressor object.

Keywords

package::AbstractString="DecisionTree": Package that provides DecisionTreeRegressor, which now has ["DecisionTree", "BetaML"] to choose from

kwargs: Valid hyperparameters of DecisionTreeRegressor

SWCForecast.manyscript — Method

manyscript(path_to_script0, ind_expr_newstrs::Pair...; append_info = false) creates a new script where the patterns in target lines are replaced by new Strings or SubstitutionStrings.

This function never overwrite old files; see also pathnorepeat.

If append_info = true, summary of the modified changes will be appended at the end of the new file.

Example

for mm = 1:12 # test training/predicting timescale
    ind_expr_newstr = [66 => r"\d+" => "1.0mm",
                        81 => r"tpast = \d+" => "tpast = 12",
                        81 => r"tfuture = \d+" => "tfuture = 2",
                        All() => r"includet" => "include"] # replace all "includet" by "include". The regular expression can match text across lines.
    manyscript(path_to_script0, ind_expr_newstr...)
end

Tips and Tricks

Comment or uncomment a certain line

ind_expr_newstr = [
    66 => r"^#\s*" => "", # uncomment line 66
    77 => r"(^.)" => s"# \1", # comment line 77
    All() => r"includet" => "include" # replace all "includet" by "include". In this case the regular expression matches text across lines.
    ]
manyscript(path_to_script0, ind_expr_newstr...)

Also see reline! and replace.

Use regular expression to find the line for replacement

For ind_expr_newstr pair, if ind is regular expression, then it searches the only line with that matches the pattern. e.g.,

ind_expr_newstr = [All() => r"includet" => "include",
                r"# run month by month" => r"^#\s*" => "", # uncommentat the line
                r"# run month by month" => r"\d+" => "1.0mm" # `reline!` on the line having the comment "# run month by month"
                ]
manyscript(path_to_script0, ind_expr_newstr...)

SWCForecast.markdownreport — Method

markdownreport(DDT::AbstractDescription) Literate the corresponding template in mdtemplatedir() generating the markdown report with all variables (informations) loaded from DDT.description_dir("description.toml"). It automatically loads "tmpl_$typeofDDT.jl" where typeofDDT is the type of DDT.

Also see description, descrption!, readdescription, writedescription for processing "description.toml".

SWCForecast.markdownreport — Method

markdownreport(path2toml, path2template, mdreport_dir) Literate the corresponding template path2template generating the markdown report with all variables (informations) loaded from path2toml. The output markdown file is saved in mdreport_dir.

In your template (e.g., "tmpl_XXXXXX.jl" in mdtemplatedir), string "PATH2TOML" will be replaced by the variable path2toml.

SWCForecast.markdownreport — Method

Example

templatename = "tmpl_DescriptOneTree.jl"
result_dir = "/home/jovyan/swc-forecast-insider/training/decisiontree_20220330/my_result_0000"
markdownreport(templatename, result_dir; findin=r"RESULT")

SWCForecast.mdimgpath! — Method

mdimgpath!(mdimg::Markdown.Image, dirnm) modify mdimg.url that mdimg.url = joinpath(dirnm, mdimg.url).

SWCForecast.mdimgpath! — Method

mdimgpath!(mdimg::Markdown.Image, dirfun::Function) modify mdimg.url that mdimg.url = dirfun(mdimg.url).

Example

thatresultdir(args...) = joinpath("foo", "bar", args...)
md1 = md"""
# Hello
this is a image:
![](foobar.png)
"""

mdimgpath!(md1.content, thatresultdir)

returns

md1 = md"""
# Hello
this is a image:
![](foo/bar/foobar.png)
"""

SWCForecast.mdimgpath! — Method

mdimgpath!(md::Markdown.MD, dirnm) = mdimgpath!(md.content, dirnm)

SWCForecast.mdimgpath! — Method

mdimgpath!(smd::SubMD, dirnm) = mdimgpath!(smd.md.content, dirnm)

SWCForecast.mdimgpath! — Method

Given a vector of Markdown contents mdc, mdimgpath!(mdc::Vector, dirnm) recursively searches Markdown.Image object and join the url (Markdown.Image.url) with dirnm (e.g., ).

Example

Before:

6-element Vector{Any}:
 Markdown.Header{2}(Any["Result"])
 Markdown.Header{3}(Any["Performance－Tree Depth"])
 Markdown.Paragraph(Any[Markdown.Image("tuned_max_depth.png", "")])
 Markdown.Header{3}(Any["Predict Result"])
 Markdown.List(Any[Any[Markdown.Paragraph(Any["Best maximum depth is 33."])]], -1, false)
 Markdown.Paragraph(Any[Markdown.Image("predict_result.png", "")])

Change the url:

julia> mdimgpath!(md.content, "RESULT_ExpeDeci_0x15b03b01222fec37")

After:

julia> md.content
6-element Vector{Any}:
 Markdown.Header{2}(Any["Result"])
 Markdown.Header{3}(Any["Performance－Tree Depth"])
 Markdown.Paragraph(Any[Markdown.Image("RESULT_ExpeDeci_0x15b03b01222fec37/tuned_max_depth.png", "")])
 Markdown.Header{3}(Any["Predict Result"])
 Markdown.List(Any[Any[Markdown.Paragraph(Any["Best maximum depth is 33."])]], -1, false)
 Markdown.Paragraph(Any[Markdown.Image("RESULT_ExpeDeci_0x15b03b01222fec37/predict_result.png", "")])

SWCForecast.mdpaths — Method

mdpaths(fnameexpr, fldrexpr::Regex, this_dir) returns a list of paths to the markdown files who matches fnameexpr. It searches only the markdown files in the folders in this_dir whose name matches fldrexpr.

SWCForecast.mdpaths — Method

mdpaths(fnameexpr::Regex, strvec::Vector, this_dir::AbstractString) is a shorthand to get the paths to target markdown files.

Example

this_dir= "/training/decisiontree_20220309/"
fnameexpr = r"brief\_report\.md"
strvec = ["ac23", "7df6", "ddf", "c37"]

mdpaths(fnameexpr, strvec, this_dir)

which returns

4-element Vector{String}:
    "/training/decisiontree_20220309/RESULT_ExpeDeci_0x15b03b01222fec37/brief_report.md"
    "/training/decisiontree_20220309/RESULT_ExpeDeci_0x8a37ffc85c22eddf/brief_report.md"
    "/training/decisiontree_20220309/RESULT_ExpeDeci_0xa14053903286ac23/brief_report.md"
    "/training/decisiontree_20220309/RESULT_ExpeDeci_0xd8a362b941617df6/brief_report.md"

SWCForecast.mdreport — Method

(DEPRECATED: use generatemyreport instead)

mdreport(DDT::DescriptOneTree, scriptpath) gives the report in juliamarkdown for script (scriptpath) based on results (DDT.description). The output folder is DDT.description_dir("brief_report.md").

SWCForecast.mkresultdir — Method

Given the path to the script file exp_file, mkresultdir return the path to the new directory named with the hash value of the file. That is, once exp_file has been changed, a new directory will be created with its path returned.

SWCForecast.mvmean — Method

mvmean(arr, n) retruns an array of element length(arr) - n of moving averaged results.

This is the function that performs best on https://stackoverflow.com/questions/59562325/moving-average-in-julia.

SWCForecast.mvnanmean — Method

mvnanmean(arr, n) use mvmean but ignoring NaN. The output array has the same dimensions as the input one, with the first n - 1 element be NaN.

WARNING

You may have to do imputation first before calculating moving average because ALL NaN are considered to be 0 when calculating moving average; that is, the average will be zero when all elements in the moving window are all NaN.

SWCForecast.mycolor_makie_palette — Method

mycolor_makie_palette()

SWCForecast.mycolor_plots_palette — Method

Example:

mycolor_plots_palette(:tab10, 5) which is equivalently Plots.palette(:tab10, 5)

See the following for color schemes:

https://docs.juliahub.com/MakieGallery/Ql23q/0.2.17/generated/colors.html

https://docs.juliaplots.org/latest/generated/colorschemes/

SWCForecast.mymovwinmean — Method

my moving average function of poor performance

SWCForecast.myslideshow — Method

Since currently Remark.slideshow has some problem:

if documenter=false, code won't be executed as expected

if documenter=true, some strings in math equations will be misreplaced

myslideshow(presentation_dir; options=Dict(), title="Title") does the following:

run Remark.slideshow()

Literate.markdown with flavor=Literate.CommonMarkFlavor(), execute=true, mdstrings=true, postprocess=hide_section

create index.html again, while other files in "build" left untouched.

Example

myslideshow(presentation_dir; options = Dict("ratio" => "16:9"))

Tips and Tricks

you can use #hide-below and #hide-above to hide the entire section

SWCForecast.narrow_types! — Method

narrow_types!(df)

Narrows the eltype of each column to the type that actually exists in the each column of dataframe.

Arguments

df: Dataframe for which you want to narrow the eltype of each column

References

https://discourse.julialang.org/t/how-to-change-field-names-and-types-of-a-dataframe/43991/9

SWCForecast.onlycodelines — Method

onlycodelines(script::AbstractString) or onlycodelines(scriptfile::AFile) returns lines that are not commented. See also iscommented

SWCForecast.print2string — Method

Praise master ZK! print2string(printfun) print content to memory and take it out as a String.

Example

printfun() = MLJDecisionTreeInterface.DT.print_tree(
    mach.fitresult.fitresult; feature_names=names(mach.fitresult.data[1]))

print2string(printfun)

SWCForecast.quantileband! — Method

quantileband!(ax, xhat, ys, percs::Vector{<:Real}, bandcolormap::ColorSchemes.ColorScheme), where colors are automatically picked from bandcolormap according to percs. It returns (percs, bds) where bd = band!(...).

Noted that percs will be sorted that the order does not matter.

SWCForecast.quantileband! — Method

quantileband!(ax, xhat::Vector{T}, ys::Vector{Vector{T}}, perc::Real, bandcolor) where T<:Real draw a band of perc quantile to ax. The band covers perc of data where 0 ≤ perc ≤ 1.

SWCForecast.quickdataoverview — Method

quickdataoverview(df_all::DataFrame; gridwidth=Month(1), xname=:datetime, figsize=(23cm, 12cm))

SWCForecast.readdescription! — Method

readdescription!(D::AbstractDescription) read "description.toml" file and return the nested dictionary (i.e., D.description in writedescription); D.description will be overwritten by the loaded one.

SWCForecast.readdescription — Method

readdescription(path2toml::AbstractString) read "description.toml" file and return the nested dictionary (i.e., D.description in writedescription).

SWCForecast.recursive_merge! — Method

See recursive_merge.

SWCForecast.recursive_merge — Method

recursive_merge merges dictionaries combining thier unique key-value pairs together. Similar to merge(d, d1, d2,...), but recursive_merge merge the dictionary recursively that sub-dictionaries are also merged, while merge(d, d1) replace the sub-dict in d by that in d1. This piece of code originally came from this thread: https://discourse.julialang.org/t/multi-layer-dict-merge/27261/2

SWCForecast.reduce_feature — Method

reduce_feature(feats0) returns unique feature names and time tags that are split by split_time_tag.

SWCForecast.reducetype — Method

reducetype(v::AbstractArray) returns the array of type Union{uniqueTYPE}, where uniqueTYPE = v .|> typeof |> unique.

For more about manipulating datatype, also see nonmissingtype in missing.jl and Base.typesplit in promotion.jl

SWCForecast.refbuildpath — Method

Given abspath, refbuildpath(abspath) returns referenced path relative to the .../build/aseets folder.

Example

julia> abspath = "/home/jovyan/swc-forecast-insider/temp_ppt/build/assets/img/trees_subset_1/tree_root_1.png"

julia> refbuildpath(abspath)
"assets/img/trees_subset_1/tree_root_1.png"

SWCForecast.reline! — Method

reline!(lines, ind::All, expr, newstr::AbstractString) replace the pattern matched by expr with newstr in the scope of fullscript = join(lines, " ").

SWCForecast.reline! — Method

reline!(lines, ind::Int, expr::Union{AbstractString, Regex}, newstr::AbstractString) replace string that matches expr in lines[ind] by newstr. newstr can be SubstitutionString. See the documentation of replace and SubstitutionString.

WARNING: if ind is an integer or ind = All(), no warning or error pops if expr matches nothing such that nothing had changed. Use reline!(lines, expr_findline::Regex, expr_replacecode::Regex, newstr::AbstractString) if you want a warning or an error for non-exclusive matching.

SWCForecast.reline! — Method

reline!(lines, expr_findline::Regex, expr_replacecode::Union{AbstractString, Regex}, newstr::AbstractString) replace string that matches expr_replacecode in lines[ind] by newstr, where ind indicates the line whose appending comment matches expr_findline. newstr can be SubstitutionString. See the documentation of replace and SubstitutionString.

SWCForecast.replacerewrite — Method

replace certain keyword in the entire script

SWCForecast.result_folder_expr — Method

result_folder_expr(strvec::Vector{<:AbstractString}) returns a single regular expression that should capture the directories for machine learning results. For example, result_folder_expr(["ac43"]) "RESULT_...0x..."

SWCForecast.returnchild — Method

returnchild(obj::Markdown.Paragraph): Given a Markdown.Paragraph, return the vector obj.text.

SWCForecast.returnchild — Method

returnchild(obj): Given an object obj::Union{Markdown.Italic, Markdown.Bold, Markdown.Header}, return the vector obj.text.

SWCForecast.runallscript — Method

Noted that the script to include shares the scope of SWCForecast module.

SWCForecast.secondyaxis — Method

secondyaxis(f_grid; color = :black, ylabel = "y") returns an Axis object of y axis to the right-hand side, with default settings for yyplot.

Example

f = Figure()
ax_right = secondyaxis(f[1,1]; color=:red)

that you can do lines!(ax_right, ...)

Further reading

Is it possible to add a secondary y axis?

SWCForecast.selectname — Method

selectname(df::DataFrame, arg) returns a vector of column names that match arg of the dataframe df. arg can be Regex, AbstractString, or InvertedIndex.

SWCForecast.selectnames — Method

selectnames(dfnames, args...) select an array of names that contains arg::String or pattern that matches arg::Regex.

Example

julia> selectnames(["Soil_water_content_10cm, water_level_#1, water_level_#2"], r"water_lev")

[water_level_#1, water_level_#2"]

SWCForecast.series2supervised — Method

To transform a time series dataset into a supervised learning dataset

Example

A = randn(500,20)
df = DataFrame(A, :auto)
X0,y0 = series_to_supervised(df[:,1:end-1], df[:,end])
X1,y1 = series2supervised(
    df[:,1:end-1] => range(-6, -1; step=1),
    df[:,end] => range(0, 0; step=-1)
    )

NOTICE!

The input DataFrame (df) must have complete rows; that is, the corresponding time tag (it might be df.datetime for example) must be consecutive because df is converted to Matrix and shifted using lag.

References:

https://machinelearningmastery.com/convert-time-series-supervised-learning-problem-python/

TODO: write test for series2supervised, by making sure the datetime shift is correct (e.g., "datetimet0" should always be 1 hour ahead of "datetimet-6" for a 10-minute sampling data).

SWCForecast.shrinkylim! — Method

shrinkylim!(ax, ydatas::Vector{<:Number}; extent = 0.15) fit y-limits to the input ydatas to with a certain extent.

SWCForecast.split_time_tag — Method

split_time_tag(str) split the feature name into variable name and the time-shift tag (which should be at the last).

Example

julia> split_time_tag("Soil_water_content_10cm_t-24")
("Soil_water_content_10cm", "t-24")

SWCForecast.symmquantile — Method

Get symmetric quantile. E.g., symmquantile(y, 0.15) returns quantile(y, [0.15, 0.85])

SWCForecast.targetrange — Method

Given a Vector, targetrange(mdcs::Vector, nlevel, exprh::Regex) find the target Markdown.Header{nlevel} object whose content matches exprh, returning a range which starts from this header to the next header Markdown.Header{nlevelnext} where nlevelnext ≤ nlevel.

Also see islevel, islevelleq and targetsection.

SWCForecast.targetsection — Method

Given a Markdown.MD object, targetsection(md1::Markdown.MD, nlevel, exprh) returns the section (which is a Vector) that starts with the Markdown.Header{nlevel} object whose content matches exprh and ends until the next header Markdown.Header{nlevelnext} where nlevelnext ≤ nlevel.

Also see targetrange.

SWCForecast.toordinal — Method

toordinal(v...) does toordinal(DateTime(v...)).

SWCForecast.toordinal — Method

toordinal(dt::DateTime)converts a Julia DateTime to a python style DateNumber. Also see datenum.

SWCForecast.tree_structure — Method

Consider deprecate this function

SWCForecast.tree_structure — Method

Given the machine mach, print the structure of tree to the output variable tree_info::string using MLJDecisionTreeInterface.DT.print_tree.

Example

tree_info = tree_structure(mach)

Consider deprecate this function

SWCForecast.tree_structure — Method

WARNING

Your pipeline must have the field selector = FeatureSelector(), e.g.,

mypipe = Pipeline(
    selector = FeatureSelector(),
    model = model
)

tuned_model = TunedModel(;
    model=mypipe)

mach = machine(
    tuned_model,
    X, # a dataframe
    y, # a vector
)

mach_report = report(mach)

SWCForecast.treeinspect! — Method

treeinspect!(FC::FeatureCounts, WEE1::DecisionTree.Root) recursively obtain the feature names and splitting thresholds of the tree nodes. Also see FeatureCounts.

Example

First, obtain the DecisionTree.Root object from the trained machine containing model of decision tree.

ftpr = fitted_params(mach1)
transformers = [v for (k, v) in ftpr.fitted_params_given_machine]
featnames = transformers[1].features_to_keep .|> string
WE = transformers[2].fitresult # WrappedEnsemble
WEE1 = WE.ensemble[1]

WARNING: how the DecisionTree.Root object WEE1 depends on the structure of your composite model.

Second, put the DecisionTree.Root object into treeinspect!

FC = FeatureCounts(4) # to at most 4th level
treeinspect!(FC, WEE1)

In final (optional), output the table

julia> DataFrame(FC; feature_names=featnames)
13×3 DataFrame
 Row │ feature                  value    level
     │ Any                      Any      Any
─────┼─────────────────────────────────────────
   1 │ precipitation_1d_t0      79.3     1
   2 │ precipitation_3d_t-2     4.9      2
   3 │ precipitation_12hr_t0    47.25    2
   4 │ humidity_CWB_t-6         81.5009  3
   5 │ precipitation_1d_t0      7.75     3
   6 │ humidity_CWB_t-24        94.2036  3
   7 │ precipitation_1hr_t-2    2.0      3
   8 │ precipitation_3d_t-24    0.65     4
   9 │ air_temperature_t-4      13.703   4
  10 │ precipitation_1d_t-18    1.4      4
  11 │ precipitation_12hr_t0    20.25    4
  12 │ precipitation_12hr_t-12  49.5     4
  13 │ precipitation_1hr_t0     2.0      4

SWCForecast.writedescription — Method

writedescription(D::AbstractDescription) create file and print D.description to "description.toml".